extracted common base model to a different file for better inheritance.
Previously, all major data-holding classes (e.g. `Event`, `User`) had to inherit the
unconfigured `BaseModel` class. This made for some smelly code repetition. Now the major
data-holding classes inherit from a single extended version of `BaseModel` that allows for
a lot cleaner code.
